pyproject.toml
[tool.poetry]
name = "forextrader"
version = "0.1.0"
description = "A trader for the forex market."
authors = ["Bart van Woesik <bartwoesik1@gmail.com>"]
readme = "README.md"
packages = [
    { include = "forex_trader" },
]

[tool.poetry.dependencies]
python = "^3.11"
scikit-learn = "^1.4.2"
pandas = "^1.2.2"
mlflow = "^2.12.1"
matplotlib = "^3.8.4"
hydra-core = "^1.3.2"
model-forge = {git = "https://github.com/BartvanWoesik/model-forge.git"}
mylogger = {git = "https://github.com/BartvanWoesik/MyLogger.git"}
seaborn = "^0.13.2"


[tool.poetry.group.dev.dependencies]
mkdocs = "^1.6.0"
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
